pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mflux-fasthtml"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
keywords = ["mflux", "flux", "flux.1", "image", "generator", "genai", "img2img", "mlx", "web", "gui", "ui"]
authors = [
  { name = "Anthony Wu", email = "please+file+gh+issue@users.noreply.github.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  "mflux>=0.4.1,<1.0",
  "python-fasthtml>=0.9.0,<1.0",
  "fastcore>=1.7,<2.0",
  "click",
  "uuid7",
]

[project.urls]
Documentation = "https://github.com/anthonywu/mflux-fasthtml#readme"
Issues = "https://github.com/anthonywu/mflux-fasthtml/issues"
Source = "https://github.com/anthonywu/mflux-fasthtml"

[project.scripts]
mflux-fasthtml-app = "mflux_fasthtml.app.main:run"  # fixme: does not start server for now

[tool.hatch.version]
path = "src/mflux_fasthtml/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.13.0,<2.0",
  "pytest>=8.3,<9.0"
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/mflux_fasthtml tests}"

[tool.coverage.run]
source_pkgs = ["mflux_fasthtml", "tests"]
branch = true
parallel = true
omit = [
  "src/mflux_fasthtml/__about__.py",
]

[tool.coverage.paths]
mflux_fasthtml = ["src/mflux_fasthtml", "*/mflux-fasthtml/src/mflux_fasthtml"]
tests = ["tests", "*/mflux-fasthtml/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`) codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["BLE", "E4", "E7", "E9", "F", "I", "ICN", "LOG", "PERF", "W"]
ignore = [
  "F405", # ignore star imports
]
